feat: set nav togglebutton render only on csr
